+++ src/main/java/com/uet/k62/web/system/examination/repository/ExamResultRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface ExamResultRepository extends JpaRepository<ExamResult, BigInteger> {
+public interface ExamResultRepository extends JpaRepository<ExamResult, Integer> {
 }
