+++ src/main/java/com/uet/k62/web/system/examination/repository/QuestionRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface QuestionRepository extends JpaRepository<Question, BigInteger> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
     List<Question> findAllByDeletedIsFalse();
-    Question findOneByIdAndDeletedIsFalse(BigInteger id);
+    Question findOneByIdAndDeletedIsFalse(Integer id);
 }
