+++ src/main/java/com/uet/k62/web/system/examination/repository/QuestionTypeRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface QuestionTypeRepository extends JpaRepository<QuestionType, BigInteger> {
+public interface QuestionTypeRepository extends JpaRepository<QuestionType, Integer> {
     List<QuestionType> findAllByDeletedIsFalse();
     QuestionType findByIdAndDeletedIsFalse(int id);
     QuestionType findOneByCode(String code);
-    QuestionType findOneByIdAndDeletedIsFalse(BigInteger id);
+    QuestionType findOneByIdAndDeletedIsFalse(Integer id);
 }
