+++ src/main/java/com/uet/k62/web/system/examination/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigInteger;
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 	List<User> findAllByDeletedIsFalse();
 	User findByUsername(String username);
-	User findUserById(BigInteger id);
-	User findByIdAndDeletedIsFalse(BigInteger id);
+	User findUserById(Integer id);
+	User findByIdAndDeletedIsFalse(Integer id);
 }
